requirements.txt
python-decouple
mysql-connector-python
customtkinter
Pillow